.raw width>=65536 support by Georgi Georgiev <chutz@chubaka.homeip.net>
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@5929 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libvo/font_load.h
@@ -16,7 +16,7 @@ typedef struct {
     raw_file* pic_a[16];
     raw_file* pic_b[16];
     short font[65536];
-    short start[65536];
+    int start[65536];   // short is not enough for unicode fonts
     short width[65536];
 } font_desc_t;
